# openapi/api_persons.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * user_authorization service - OpenAPI 3.0
 *
 * Сервис авторизации пользователей
 *
 * API version: 1.0.0
 */

package openapi

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// PersonsAPIController binds http requests to an api service and writes the service results to the http response
type PersonsAPIController struct {
	service      PersonsAPIServicer
	errorHandler ErrorHandler
}

// PersonsAPIOption for how the controller is set up.
type PersonsAPIOption func(*PersonsAPIController)

// WithPersonsAPIErrorHandler inject ErrorHandler into controller
func WithPersonsAPIErrorHandler(h ErrorHandler) PersonsAPIOption {
	return func(c *PersonsAPIController) {
		c.errorHandler = h
	}
}

// NewPersonsAPIController creates a default api controller
func NewPersonsAPIController(s PersonsAPIServicer, opts ...PersonsAPIOption) *PersonsAPIController {
	controller := &PersonsAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the PersonsAPIController
func (c *PersonsAPIController) Routes() Routes {
	return Routes{
		"PersonsList": Route{
			strings.ToUpper("Get"),
			"/api/v1/persons",
			c.PersonsList,
		},
		"GetPersonByUUID": Route{
			strings.ToUpper("Get"),
			"/api/v1/persons/{uuid}",
			c.GetPersonByUUID,
		},
		"ChangePersonByUUUID": Route{
			strings.ToUpper("Put"),
			"/api/v1/persons/{uuid}",
			c.ChangePersonByUUUID,
		},
		"DeletePersonByUUUID": Route{
			strings.ToUpper("Delete"),
			"/api/v1/persons/{uuid}",
			c.DeletePersonByUUUID,
		},
	}
}

// PersonsList - Persons list
func (c *PersonsAPIController) PersonsList(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.PersonsList(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetPersonByUUID - Find person by uuid
func (c *PersonsAPIController) GetPersonByUUID(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	uuidParam := params["uuid"]
	if uuidParam == "" {
		c.errorHandler(w, r, &RequiredError{"uuid"}, nil)
		return
	}
	result, err := c.service.GetPersonByUUID(r.Context(), uuidParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// ChangePersonByUUUID - change person by uuid
func (c *PersonsAPIController) ChangePersonByUUUID(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	uuidParam := params["uuid"]
	if uuidParam == "" {
		c.errorHandler(w, r, &RequiredError{"uuid"}, nil)
		return
	}
	personParam := Person{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&personParam); err != nil && !errors.Is(err, io.EOF) {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertPersonRequired(personParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertPersonConstraints(personParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.ChangePersonByUUUID(r.Context(), uuidParam, personParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeletePersonByUUUID - delete person by uuid
func (c *PersonsAPIController) DeletePersonByUUUID(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	uuidParam := params["uuid"]
	if uuidParam == "" {
		c.errorHandler(w, r, &RequiredError{"uuid"}, nil)
		return
	}
	result, err := c.service.DeletePersonByUUUID(r.Context(), uuidParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}